Reorganizes file structure for components
